Model migrations and stuff

+++ Ifc/src/org/bimserver/ifc/IfcModel.java
@@ -306,7 +306,7 @@ public class IfcModel implements IfcModelInterface {
 		return nameIndex.get(eClass).get(name);
 	}
 
-	public long size() {
+	public int size() {
 		return objects.size() + unidentifiedObjects.size();
 	}
